Possible fix for Bug #37797 (cherry picked from commit b882208caf95bf931409093d22a45c4ad262dc70)
Conflicts:
	com.openexchange.calendar/src/com/openexchange/calendar/CalendarMySQL.java
	com.openexchange.calendar/src/com/openexchange/calendar/recurrence/RecurringCalculation.java

+++ com.openexchange.calendar/src/com/openexchange/calendar/CalendarMySQL.java
@@ -2761,8 +2761,13 @@ public class CalendarMySQL implements CalendarSqlImp {
                 throw OXCalendarExceptionCodes.INVALID_RECURRENCE_TYPE_CHANGE.create(new Object[0]);
             }
             if (cdao.getRecurrenceType() == CalendarDataObject.NO_RECURRENCE) {
-                cdao.removeRecurrenceID();
-                cdao.removeRecurrencePosition();
+                if (cdao.containsRecurrenceType()) {
+                    cdao.setRecurrenceID(0);
+                    cdao.setRecurrencePosition(-1);
+                } else {
+                    cdao.removeRecurrenceID();
+                    cdao.removeRecurrencePosition();
+                }
             }
             final List<Integer> exceptions = getExceptionList(null, ctx, edao.getRecurrenceID());
             if (exceptions != null && !exceptions.isEmpty()) {
